Address Copilot review on get_combined_metadata
- Widen thresholds: int | list[int] | None -> float | list[float] | None
  to match the docstring's "numeric or list of numbers" promise. The
  Water Data API treats threshold values as floats, so the previous
  int-only annotation was misleading downstream type-checked callers.
- In test_get_combined_metadata_multi_site_post, swap the unused `md`
  binding for `_` to match the convention used by the other live
  waterdata tests in this file (`df, _ = get_*(...)`). The companion
  test_get_combined_metadata still binds `md` because it asserts on
  metadata attributes.

+++ dataretrieval/waterdata/api.py
@@ -941,7 +941,7 @@ def get_combined_metadata(
     statistic_id: str | list[str] | None = None,
     data_type: str | list[str] | None = None,
     computation_identifier: str | list[str] | None = None,
-    thresholds: int | list[int] | None = None,
+    thresholds: float | list[float] | None = None,
     sublocation_identifier: str | list[str] | None = None,
     primary: str | list[str] | None = None,
     parent_time_series_id: str | list[str] | None = None,
